refactor: update layouts
