+++ src/index.css
@@ -1,4 +1,21 @@
 .container {
   width: min(90%, 1280px);
-  margin-inline: start;
+  margin-inline: auto;
+}
+
+.flex {
+  display: flex;
+}
+
+.items-center {
+  align-items: center;
+}
+
+.gap-x-3 {
+  column-gap: 0.75rem;
+}
+
+.py-4 {
+  padding-top: 1rem;
+  padding-bottom: 1rem;
 }
