+++ src/main.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 
 import { RouterProvider } from 'react-router-dom';
-import { router } from './router';
 
-import './index.css';
+import { router } from '~/router';
+import '~/index.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
