+++ src/pages/index.ts
@@ -1,3 +1,3 @@
-export { default as Home } from './Home/Home';
-export { default as About } from './About/About';
-export { default as NotFound } from './NotFound/NotFound';
+export { default as Home } from './home/Home';
+export { default as About } from './about/About';
+export { default as NotFound } from './not-found/NotFound';
